Update input max_output_boxes_per_class and output selected_indices to type int64

+++ onnx/backend/test/case/node/nonmaxsuppression.py
@@ -28,10 +28,10 @@ class NonMaxSuppression(Base):
             [0.0, 100.0, 1.0, 101.0]
         ]]).astype(np.float32)
         scores = np.array([[[0.9, 0.75, 0.6, 0.95, 0.5, 0.3]]]).astype(np.float32)
-        max_output_boxes_per_class = np.array([3]).astype(np.int32)
+        max_output_boxes_per_class = np.array([3]).astype(np.int64)
         iou_threshold = np.array([0.5]).astype(np.float32)
         score_threshold = np.array([0.0]).astype(np.float32)
-        selected_indices = np.array([[0, 0, 3], [0, 0, 0], [0, 0, 5]]).astype(np.int32)
+        selected_indices = np.array([[0, 0, 3], [0, 0, 0], [0, 0, 5]]).astype(np.int64)
 
         expect(node, inputs=[boxes, scores, max_output_boxes_per_class, iou_threshold, score_threshold], outputs=[selected_indices], name='test_nonmaxsuppression_suppress_by_IOU')
 
@@ -51,10 +51,10 @@ class NonMaxSuppression(Base):
             [0.0, 100.0, 1.0, 101.0]
         ]]).astype(np.float32)
         scores = np.array([[[0.9, 0.75, 0.6, 0.95, 0.5, 0.3]]]).astype(np.float32)
-        max_output_boxes_per_class = np.array([3]).astype(np.int32)
+        max_output_boxes_per_class = np.array([3]).astype(np.int64)
         iou_threshold = np.array([0.5]).astype(np.float32)
         score_threshold = np.array([0.4]).astype(np.float32)
-        selected_indices = np.array([[0, 0, 3], [0, 0, 0]]).astype(np.int32)
+        selected_indices = np.array([[0, 0, 3], [0, 0, 0]]).astype(np.int64)
 
         expect(node, inputs=[boxes, scores, max_output_boxes_per_class, iou_threshold, score_threshold], outputs=[selected_indices], name='test_nonmaxsuppression_suppress_by_IOU_and_scores')
 
@@ -74,10 +74,10 @@ class NonMaxSuppression(Base):
             [1.0, 101.0, 0.0, 100.0]
         ]]).astype(np.float32)
         scores = np.array([[[0.9, 0.75, 0.6, 0.95, 0.5, 0.3]]]).astype(np.float32)
-        max_output_boxes_per_class = np.array([3]).astype(np.int32)
+        max_output_boxes_per_class = np.array([3]).astype(np.int64)
         iou_threshold = np.array([0.5]).astype(np.float32)
         score_threshold = np.array([0.0]).astype(np.float32)
-        selected_indices = np.array([[0, 0, 3], [0, 0, 0], [0, 0, 5]]).astype(np.int32)
+        selected_indices = np.array([[0, 0, 3], [0, 0, 0], [0, 0, 5]]).astype(np.int64)
 
         expect(node, inputs=[boxes, scores, max_output_boxes_per_class, iou_threshold, score_threshold], outputs=[selected_indices], name='test_nonmaxsuppression_flipped_coordinates')
 
@@ -97,10 +97,10 @@ class NonMaxSuppression(Base):
             [0.0, 100.0, 1.0, 101.0]
         ]]).astype(np.float32)
         scores = np.array([[[0.9, 0.75, 0.6, 0.95, 0.5, 0.3]]]).astype(np.float32)
-        max_output_boxes_per_class = np.array([2]).astype(np.int32)
+        max_output_boxes_per_class = np.array([2]).astype(np.int64)
         iou_threshold = np.array([0.5]).astype(np.float32)
         score_threshold = np.array([0.0]).astype(np.float32)
-        selected_indices = np.array([[0, 0, 3], [0, 0, 0]]).astype(np.int32)
+        selected_indices = np.array([[0, 0, 3], [0, 0, 0]]).astype(np.int64)
 
         expect(node, inputs=[boxes, scores, max_output_boxes_per_class, iou_threshold, score_threshold], outputs=[selected_indices], name='test_nonmaxsuppression_limit_output_size')
 
@@ -115,10 +115,10 @@ class NonMaxSuppression(Base):
             [0.0, 0.0, 1.0, 1.0]
         ]]).astype(np.float32)
         scores = np.array([[[0.9]]]).astype(np.float32)
-        max_output_boxes_per_class = np.array([3]).astype(np.int32)
+        max_output_boxes_per_class = np.array([3]).astype(np.int64)
         iou_threshold = np.array([0.5]).astype(np.float32)
         score_threshold = np.array([0.0]).astype(np.float32)
-        selected_indices = np.array([[0, 0, 0]]).astype(np.int32)
+        selected_indices = np.array([[0, 0, 0]]).astype(np.int64)
 
         expect(node, inputs=[boxes, scores, max_output_boxes_per_class, iou_threshold, score_threshold], outputs=[selected_indices], name='test_nonmaxsuppression_single_box')
 
@@ -143,10 +143,10 @@ class NonMaxSuppression(Base):
             [0.0, 0.0, 1.0, 1.0]
         ]]).astype(np.float32)
         scores = np.array([[[0.9, 0.9, 0.9, 0.9, 0.9, 0.9, 0.9, 0.9, 0.9, 0.9]]]).astype(np.float32)
-        max_output_boxes_per_class = np.array([3]).astype(np.int32)
+        max_output_boxes_per_class = np.array([3]).astype(np.int64)
         iou_threshold = np.array([0.5]).astype(np.float32)
         score_threshold = np.array([0.0]).astype(np.float32)
-        selected_indices = np.array([[0, 0, 0]]).astype(np.int32)
+        selected_indices = np.array([[0, 0, 0]]).astype(np.int64)
 
         expect(node, inputs=[boxes, scores, max_output_boxes_per_class, iou_threshold, score_threshold], outputs=[selected_indices], name='test_nonmaxsuppression_identical_boxes')
 
@@ -167,10 +167,10 @@ class NonMaxSuppression(Base):
             [0.5, 100.5, 1.0, 1.0]
         ]]).astype(np.float32)
         scores = np.array([[[0.9, 0.75, 0.6, 0.95, 0.5, 0.3]]]).astype(np.float32)
-        max_output_boxes_per_class = np.array([3]).astype(np.int32)
+        max_output_boxes_per_class = np.array([3]).astype(np.int64)
         iou_threshold = np.array([0.5]).astype(np.float32)
         score_threshold = np.array([0.0]).astype(np.float32)
-        selected_indices = np.array([[0, 0, 3], [0, 0, 0], [0, 0, 5]]).astype(np.int32)
+        selected_indices = np.array([[0, 0, 3], [0, 0, 0], [0, 0, 5]]).astype(np.int64)
 
         expect(node, inputs=[boxes, scores, max_output_boxes_per_class, iou_threshold, score_threshold], outputs=[selected_indices], name='test_nonmaxsuppression_center_point_box_format')
 
@@ -191,10 +191,10 @@ class NonMaxSuppression(Base):
         ]]).astype(np.float32)
         scores = np.array([[[0.9, 0.75, 0.6, 0.95, 0.5, 0.3],
                             [0.9, 0.75, 0.6, 0.95, 0.5, 0.3]]]).astype(np.float32)
-        max_output_boxes_per_class = np.array([2]).astype(np.int32)
+        max_output_boxes_per_class = np.array([2]).astype(np.int64)
         iou_threshold = np.array([0.5]).astype(np.float32)
         score_threshold = np.array([0.0]).astype(np.float32)
-        selected_indices = np.array([[0, 0, 3], [0, 0, 0], [0, 1, 3], [0, 1, 0]]).astype(np.int32)
+        selected_indices = np.array([[0, 0, 3], [0, 0, 0], [0, 1, 3], [0, 1, 0]]).astype(np.int64)
 
         expect(node, inputs=[boxes, scores, max_output_boxes_per_class, iou_threshold, score_threshold], outputs=[selected_indices], name='test_nonmaxsuppression_two_classes')
 
@@ -222,9 +222,9 @@ class NonMaxSuppression(Base):
         ]]).astype(np.float32)
         scores = np.array([[[0.9, 0.75, 0.6, 0.95, 0.5, 0.3],
                             [0.9, 0.75, 0.6, 0.95, 0.5, 0.3]]]).astype(np.float32)
-        max_output_boxes_per_class = np.array([2]).astype(np.int32)
+        max_output_boxes_per_class = np.array([2]).astype(np.int64)
         iou_threshold = np.array([0.5]).astype(np.float32)
         score_threshold = np.array([0.0]).astype(np.float32)
-        selected_indices = np.array([[0, 0, 3], [0, 0, 0], [1, 0, 3], [1, 0, 0]]).astype(np.int32)
+        selected_indices = np.array([[0, 0, 3], [0, 0, 0], [1, 0, 3], [1, 0, 0]]).astype(np.int64)
 
         expect(node, inputs=[boxes, scores, max_output_boxes_per_class, iou_threshold, score_threshold], outputs=[selected_indices], name='test_nonmaxsuppression_two_batches')
